requirements.txt
eth-brownie==1.12.4
munch==2.5.0
fixedint==0.1.6
helpers==0.2.0
pytest-html==3.1.1
